requirements.txt
pyrogram
tgcrypto
python-dotenv
aiohttp
FFmpeg-python
pydantic
